Finished up cables, working on refactoring
Added a new motherboard for the buildMode as the older version had a weird power connector
Added the power connector for the new motherboard
Removed several irrelevant headers(that are optional) on the motherboard.
I still need to add the hover description for the new cables and the rest of the parts on the motherboard.